Fix album shortcut to work for more photos and screens.
The various screens such as /photo, /share, /album use different menu
layouts. Also, pictures with motion will display an option to download
a video in addition to a normal download button.

This change is ugly and makes the logic dependent on potentially
locale-dependent strings. In particular, the keyboard shortcuts are
assumed to be locale-neutral (works for en, ja, ko). The pattern I saw
and implemented support for is that the menu item for adding to an
album is always after the menu item with a shortcut of shift+r (if it
exists), after the entry with the shortcut of shift+d (if it exists),
and after the second entry containing text from the shift+d entry (if
it exists). This seems to work reliably.

Fixes #1.